# api-docs.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "G-Zang (归藏) API文档",
    "description": "G-Zang (归藏) 财务管理系统API文档",
    "version": "1.0.0",
    "contact": {
      "name": "G-Zang Team",
      "email": "team@gzang.com"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "本地开发环境"
    }
  ],
  "security": [
    {
      "Bearer Authentication": []
    }
  ],
  "components": {
    "securitySchemes": {
      "Bearer Authentication": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "在请求头中添加 JWT Token：Authorization: Bearer {token}"
      }
    },
    "schemas": {
...
      "Transaction": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "交易ID"
          },
          "userId": {
            "type": "integer",
            "format": "int64",
            "description": "记录用户ID"
          },
          "companyId": {
            "type": "integer",
            "format": "int64",
            "description": "所属公司ID"
          },
          "amount": {
            "type": "number",
            "format": "decimal",
            "description": "交易金额"
          },
          "type": {
            "type": "integer",
            "description": "交易类型 (1:收入, 2:支出)"
          },
          "categoryId": {
            "type": "integer",
            "format": "int64",
            "description": "分类ID"
          },
          "accountId": {
            "type": "integer",
            "format": "int64",
            "description": "账户ID"
          },
          "transactionTime": {
            "type": "string",
            "format": "date-time",
            "description": "交易发生时间"
          },
          "remark": {
            "type": "string",
            "description": "备注"
          },
          "relatedBusinessId": {
            "type": "string",
            "description": "关联业务ID"
          },
          "createTime": {
            "type": "string",
            "format": "date-time",
            "description": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "format": "date-time",
            "description": "更新时间"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/transactions": {
      "post": {
        "tags": ["交易管理"],
        "summary": "创建交易记录",
        "description": "新增一笔交易记录",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "get": {
        "tags": ["交易管理"],
        "summary": "分页查询交易记录",
        "description": "分页获取交易记录列表",
        "parameters": [
          {
            "name": "current",
            "in": "query",
            "description": "页码",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "每页大小",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "description": "开始时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "交易类型 (1:收入, 2:支出)",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "categoryId",
            "in": "query",
            "description": "分类ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/v1/transactions/summary": {
      "get": {
        "tags": ["交易管理"],
        "summary": "获取交易汇总",
        "description": "获取交易汇总统计信息",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "description": "开始时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/v1/transactions/category-summary": {
      "get": {
        "tags": ["交易管理"],
        "summary": "分类统计",
        "description": "按分类统计交易金额",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "description": "开始时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "交易类型 (1:收入, 2:支出)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/v1/transactions/{id}": {
      "get": {
        "tags": ["交易管理"],
        "summary": "获取交易记录详情",
        "description": "根据ID获取交易记录详情",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "交易记录ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "put": {
        "tags": ["交易管理"],
        "summary": "更新交易记录",
        "description": "更新指定交易记录",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "交易记录ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Transaction"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      },
      "delete": {
        "tags": ["交易管理"],
        "summary": "删除交易记录",
        "description": "删除指定交易记录",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "交易记录ID",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/api/v1/reports/income-expense-summary": {
      "get": {
        "tags": ["报表统计"],
        "summary": "收支汇总",
        "description": "获取收入支出汇总统计",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "description": "开始时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/v1/reports/monthly-trend": {
      "get": {
        "tags": ["报表统计"],
        "summary": "月度趋势",
        "description": "获取指定年份的月度收支趋势",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "description": "年份",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
    "/api/v1/reports/category-report": {
      "get": {
        "tags": ["报表统计"],
        "summary": "分类统计",
        "description": "按分类统计收支情况",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "description": "开始时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "description": "结束时间",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "type",
            "in": "query",
            "description": "交易类型 (1:收入, 2:支出)",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    },
...
    "/api/v1/reports/yearly-comparison": {
      "get": {
        "tags": ["报表统计"],
        "summary": "年度对比",
        "description": "多年度收支对比分析",
        "parameters": [
          {
            "name": "years",
            "in": "query",
            "required": true,
            "description": "年份列表，用逗号分隔",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功"
          }
        }
      }
    }
  }
}